// app/cms/models/cms-size.ts
/**
 * Copyright (c) 2016 Barco n.v. All Rights Reserved. This software is confidential and proprietary information of Barco n.v.
 * ("Confidential Information"). You shall not disclose such Confidential Information and shall use it only in accordance with
 * the terms of the license agreement you entered into with Barco.
 */

/**
 * The interface defines the model for size of a 2D rectangular object.
 */
export interface ISize {
    width: number;
    height: number;
}